# Conflicts:
#	src/main/java/br/gov/sc/pm/emergencia/gtlock/operadoras/controller/mapper/OperadorasMapper.java

+++ src/main/java/br/gov/sc/pm/emergencia/gtlock/operadoras/controller/dto/OperadorasDTO.java
@@ -1,8 +1,8 @@
 package br.gov.sc.pm.emergencia.gtlock.operadoras.controller.dto;
 
 public record OperadorasDTO(
-        Long operadora,
-        Integer id,
+        Integer operadora,
+        Long id,
         String descricao
 ) {
 }
